requirements.txt
tensorflow-gpu==2.1.0
keras==2.3.1
pathlib2==2.3.5
numpy==1.18.1
opencv-python==4.2.0.32
pandas==0.25.1
matplotlib==3.1.1
